# Improve performance of marshalling small keys