## Why this should be merged

- This PR improves the memory usage + improves time to encode small keys.
- For larger keys (larger than 64 bytes) there may be a few ns overhead (but the profile actually came out faster when I ran it)... But I'd expect most keys to be around 32 bytes (as a hash) or less anyways.

Before:
```
Benchmark_EncodeKey/empty-12         	28405167	        38.21 ns/op	      16 B/op	       1 allocs/op
Benchmark_EncodeKey/1_byte-12        	32087634	        40.47 ns/op	      16 B/op	       1 allocs/op
Benchmark_EncodeKey/2_bytes-12       	31305470	        38.91 ns/op	      16 B/op	       1 allocs/op
Benchmark_EncodeKey/4_bytes-12       	32088528	        38.95 ns/op	      16 B/op	       1 allocs/op
Benchmark_EncodeKey/8_bytes-12       	25685365	        46.60 ns/op	      24 B/op	       1 allocs/op
Benchmark_EncodeKey/32_bytes-12      	21811918	        62.90 ns/op	      48 B/op	       1 allocs/op
Benchmark_EncodeKey/64_bytes-12      	14426572	        85.85 ns/op	      80 B/op	       1 allocs/op
Benchmark_EncodeKey/1024_bytes-12    	 1329334	       909.2  ns/op	    1152 B/op	       1 allocs/op
```

After:
```
Benchmark_EncodeKey/empty-12         	46378155	        25.56 ns/op	       1 B/op	       1 allocs/op
Benchmark_EncodeKey/1_byte-12        	43711533	        28.16 ns/op	       2 B/op	       1 allocs/op
Benchmark_EncodeKey/2_bytes-12       	41919078	        28.95 ns/op	       3 B/op	       1 allocs/op
Benchmark_EncodeKey/4_bytes-12       	42343974	        29.56 ns/op	       5 B/op	       1 allocs/op
Benchmark_EncodeKey/8_bytes-12       	30829506	        39.67 ns/op	      16 B/op	       1 allocs/op
Benchmark_EncodeKey/32_bytes-12      	20236386	        60.69 ns/op	      48 B/op	       1 allocs/op
Benchmark_EncodeKey/64_bytes-12      	14738362	        83.64 ns/op	      80 B/op	       1 allocs/op
Benchmark_EncodeKey/1024_bytes-12    	 1401087	       839.5  ns/op	    1152 B/op	       1 allocs/op
```

## How this works

We have a `uintSize` function so we don't need to estimate the length.

## How this was tested

- [X] Added unit tests
- [X] Added benchmarks
- [X] CI